+++ packages/ui-components/src/components/Map/TilePicker/index.js
@@ -5,3 +5,4 @@
  */
 
 export * from './TilePicker';
+export * from './constants';
