+++ packages/ui-components/src/components/ReferenceTooltip.js
@@ -1,38 +1,31 @@
-/**
- * Tupaia Web
- * Copyright (c) 2021 Beyond Essential Systems Pty Ltd.
- * This source code is licensed under the AGPL-3.0 license
- * found in the LICENSE file in the root directory of this source tree.
+/*
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ *
  */
-
 import React from 'react';
-
 import { Typography, Tooltip, withStyles } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import InfoRoundedIcon from '@material-ui/icons/InfoRounded';
 import styled from 'styled-components';
-import { BLUE } from '../styles';
 
 const DEFAULT = 'default';
 const TILE_SET = 'tileSet';
 const MAP_OVERLAY = 'mayOverlay';
-export const TOOLTIP_ICON_STYLE_OPTIONS = {
-  DEFAULT,
-  TILE_SET,
-  MAP_OVERLAY,
-};
 
 const IconButton = styled(InfoRoundedIcon)`
-  fontsize: 16px;
+  font-size: 16px;
   color: grey;
   transition: color 0.2s ease;
+
   &:hover {
     background-color: initial;
     color: white;
   }
 `;
+
 const styles = {
-  link: { color: BLUE },
+  link: { color: '#22c7fc' },
   iconButton: {
     [DEFAULT]: { fontSize: '16px' },
     [TILE_SET]: { fontSize: '16px', marginBottom: '-1px' },
@@ -40,6 +33,7 @@ const styles = {
   },
   typography: { backgroundColor: 'black' },
 };
+
 const StyledToolTip = withStyles(theme => ({
   arrow: {
     color: theme.palette.common.black,
@@ -73,7 +67,7 @@ export const ReferenceTooltip = props => {
   const content = getContent(reference);
 
   return (
-    <StyledToolTip arrow interactive placement="top" enterTouchDelay="50" title={content}>
+    <StyledToolTip arrow interactive placement="top" enterTouchDelay={50} title={content}>
       <IconButton
         style={
           iconStyleOption && styles.iconButton[iconStyleOption]
