--- packages/web-frontend/src/containers/MapControl/contants.js
@@ -1,12 +1,0 @@
-/*
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-
-import PropTypes from 'prop-types';
-
-export const tileSetShape = {
-  key: PropTypes.string,
-  label: PropTypes.string,
-  thumbnail: PropTypes.string,
-};
