--- packages/web-frontend/src/containers/MapControl/index.js
@@ -1,6 +1,0 @@
-/*
- * Tupaia
- * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
- */
-
-export * from './MapControl';
